task: move all methods from old to new ctx store
In this first iteration, the purpose was to remove the innovation.store
and innovation.service for the new store. The functionality was kept as
is without taking into consideration if it could or should be moved.
