+++ src/app/base/core.component.ts
@@ -15,13 +15,12 @@ import { EnvironmentVariablesStore } from '@modules/core/stores/environment-vari
 import { AuthenticationStore } from '@modules/stores/authentication/authentication.store';
 import { ContextStore } from '@modules/stores/context/context.store';
 import { ContextPageLayoutType, ContextPageStatusType } from '@modules/stores/context/context.types';
-import { InnovationStore } from '@modules/stores/innovation/innovation.store';
 
 import { AlertType, LinkType, MappedObjectType } from '@modules/core/interfaces/base.interfaces';
 import { URLS } from './constants';
 import { UtilsHelper } from './helpers';
 import { InnovationRecordSchemaStore } from '@modules/stores/innovation/innovation-record/innovation-record-schema/innovation-record-schema.store';
-import { CtxStore, InnovationContextStore } from '@modules/stores';
+import { CtxStore } from '@modules/stores';
 
 @Component({ template: '' })
 export class CoreComponent implements OnDestroy {
@@ -50,7 +49,6 @@ export class CoreComponent implements OnDestroy {
   protected stores: {
     authentication: AuthenticationStore;
     context: ContextStore;
-    innovation: InnovationStore;
     schema: InnovationRecordSchemaStore;
   };
 
@@ -94,7 +92,6 @@ export class CoreComponent implements OnDestroy {
     this.stores = {
       authentication: injector.get(AuthenticationStore),
       context: injector.get(ContextStore),
-      innovation: injector.get(InnovationStore),
       schema: injector.get(InnovationRecordSchemaStore)
     };
