+++ src/app/base/core.service.ts
@@ -10,7 +10,6 @@ import { AuthenticationStore } from '@modules/stores/authentication/authenticati
 import { UserRoleEnum } from '@modules/stores/authentication/authentication.enums';
 import { ContextStore } from '@modules/stores/context/context.store';
 import { ContextPageLayoutType } from '@modules/stores/context/context.types';
-import { InnovationStore } from '@modules/stores/innovation/innovation.store';
 import { CtxStore, InnovationRecordSchemaStore } from '@modules/stores';
 
 @Injectable()
@@ -24,7 +23,6 @@ export class CoreService {
   protected stores: {
     authentication: AuthenticationStore;
     context: ContextStore;
-    innovation: InnovationStore;
     schema: InnovationRecordSchemaStore;
   };
 
@@ -47,7 +45,6 @@ export class CoreService {
     this.stores = {
       authentication: injector.get(AuthenticationStore),
       context: injector.get(ContextStore),
-      innovation: injector.get(InnovationStore),
       schema: injector.get(InnovationRecordSchemaStore)
     };
