+++ src/modules/core/helpers/utils.helper.ts
@@ -8,7 +8,7 @@ import {
   SupportUpdatedResponseDTO
 } from '@modules/feature-modules/accessor/services/accessor.service';
 import { OrganisationsListDTO } from '@modules/shared/services/organisations.service';
-import { InnovationRecordSchemaStore, InnovationStore } from '@modules/stores';
+import { InnovationRecordSchemaStore } from '@modules/stores';
 import { PhoneUserPreferenceEnum } from '@modules/stores/authentication/authentication.service';
 
 export class UtilsHelper {
