+++ src/modules/feature-modules/accessor/pages/innovation/overview/overview.component.ts
@@ -12,9 +12,9 @@ import { InnovationCollaboratorsListDTO } from '@modules/shared/services/innovat
 import { InnovationsService } from '@modules/shared/services/innovations.service';
 import { InnovationStatisticsEnum, UserStatisticsTypeEnum } from '@modules/shared/services/statistics.enum';
 import { StatisticsService } from '@modules/shared/services/statistics.service';
-import { InnovationService } from '@modules/stores';
 import { InnovationUnitSuggestionsType } from '@modules/stores/innovation/innovation.models';
 import { KeyProgressAreasPayloadType } from '@modules/theme/components/key-progress-areas-card/key-progress-areas-card.component';
+import { InnovationContextService } from '@modules/stores';
 
 @Component({
   selector: 'app-accessor-pages-innovation-overview',
@@ -23,7 +23,7 @@ import { KeyProgressAreasPayloadType } from '@modules/theme/components/key-progr
 export class InnovationOverviewComponent extends CoreComponent implements OnInit {
   innovationId: string;
   innovation: ContextInnovationType;
-  innovationSupportStatus = this.stores.innovation.INNOVATION_SUPPORT_STATUS;
+  innovationSupportStatus = this.ctx.innovation.INNOVATION_SUPPORT_STATUS;
 
   qaSuggestions: InnovationUnitSuggestionsType = [];
 
@@ -55,7 +55,7 @@ export class InnovationOverviewComponent extends CoreComponent implements OnInit
   constructor(
     private activatedRoute: ActivatedRoute,
     private innovationsService: InnovationsService,
-    private innovationService: InnovationService,
+    private innovationService: InnovationContextService,
     private statisticsService: StatisticsService
   ) {
     super();
