+++ src/modules/shared/pages/innovation/everyone-working-on-innovation/everyone-working-on-innovation.component.spec.ts
@@ -6,15 +6,13 @@ import { ActivatedRoute, RouterModule } from '@angular/router';
 import { of, throwError } from 'rxjs';
 
 import { CoreModule, AppInjector } from '@modules/core';
-import { StoresModule, InnovationStore } from '@modules/stores';
+import { StoresModule } from '@modules/stores';
 import { SharedModule } from '@modules/shared/shared.module';
 
 import { PageEveryoneWorkingOnInnovationComponent } from './everyone-working-on-innovation.component';
 
 describe('Shared/Pages/Innovation/Messages/PageInnovationThreadMessagesListComponent', () => {
   let activatedRoute: ActivatedRoute;
-
-  let innovationStore: InnovationStore;
 
   let component: PageEveryoneWorkingOnInnovationComponent;
   let fixture: ComponentFixture<PageEveryoneWorkingOnInnovationComponent>;
@@ -27,8 +25,6 @@ describe('Shared/Pages/Innovation/Messages/PageInnovationThreadMessagesListCompo
     AppInjector.setInjector(TestBed.inject(Injector));
 
     activatedRoute = TestBed.inject(ActivatedRoute);
-
-    innovationStore = TestBed.inject(InnovationStore);
 
     activatedRoute.snapshot.data = { innovationData: { id: 'Inno01', name: 'Innovation 01', assessment: {} } };
   });
