+++ src/modules/shared/pages/innovation/record/innovation-record.component.ts
@@ -87,7 +87,7 @@ export class PageInnovationRecordComponent extends CoreComponent implements OnIn
     this.setPageTitle('Innovation record');
 
     forkJoin([
-      this.stores.innovation.getSectionsSummary$(this.activatedRoute.snapshot.params.innovationId),
+      this.ctx.innovation.getSectionsSummary$(this.activatedRoute.snapshot.params.innovationId),
       ...(this.isInnovatorType
         ? [
             this.statisticsService.getInnovationStatisticsInfo(this.innovationId, {
