+++ src/modules/shared/resolvers/innovation-section-data-v3.resolver.ts
@@ -2,13 +2,12 @@ import { inject } from '@angular/core';
 import { ActivatedRouteSnapshot, ResolveFn } from '@angular/router';
 import { Observable, of } from 'rxjs';
 
-import { InnovationRecordSchemaStore, InnovationStore } from '@modules/stores';
+import { InnovationRecordSchemaStore } from '@modules/stores';
 // import { getInnovationRecordSchemaTranslationsMap } from '@modules/stores/innovation/innovation-record/202405/ir-v3.helpers';
 
 export const innovationSectionDataResolverV3: ResolveFn<any> = (
   route: ActivatedRouteSnapshot
 ): Observable<{ id: null | string; name: string }> => {
-  const innovationStore: InnovationStore = inject(InnovationStore);
   const irSchemaStore: InnovationRecordSchemaStore = inject(InnovationRecordSchemaStore);
 
   return of({
