+++ src/modules/stores/ctx/innovation/innovation-context.service.ts
@@ -6,12 +6,25 @@ import { map, take } from 'rxjs/operators';
 import { UrlModel } from '@modules/core/models/url.model';
 import { EnvironmentVariablesStore } from '@modules/core/stores/environment-variables.store';
 
-import { AuthenticationModel } from '@modules/stores/authentication/authentication.models';
+import { AuthenticationModel } from '../../authentication/authentication.models';
 import { UserRoleEnum } from '@app/base/enums';
 import { InnovationInfoDTO } from '@modules/shared/services/innovations.dtos';
-import { InnovationStatusEnum, InnovationSupportStatusEnum } from '@modules/stores/innovation';
-import { InnovationGroupedStatusEnum } from '@modules/stores/innovation/innovation.enums';
+import {
+  InnovationGroupedStatusEnum,
+  InnovationStatusEnum,
+  InnovationSupportStatusEnum
+} from '../../innovation/innovation.enums';
 import { ContextInnovationType } from './innovation-context.types';
+import {
+  GetInnovationEvidenceDTO,
+  INNOVATION_STATUS,
+  InnovationAllSectionsInfoDTO,
+  InnovationSectionInfoDTO,
+  InnovationSectionsListDTO,
+  InnovationUnitSuggestionsType,
+  OrganisationSuggestionModel
+} from '../../innovation/innovation.models';
+import { MappedObjectType } from '@app/base/types';
 
 @Injectable()
 export class InnovationContextService {
@@ -111,5 +124,126 @@ export class InnovationContextService {
         };
       })
     );
+  }
+
+  submitInnovation(innovationId: string): Observable<{ id: string; status: keyof typeof INNOVATION_STATUS }> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/submit')
+      .setPathParams({ innovationId });
+    return this.http.patch<{ id: string; status: keyof typeof INNOVATION_STATUS }>(url.buildUrl(), {}).pipe(
+      take(1)
+      // finalize(() => this.ctx.innovation.clear()) // TODO: Check this.
+    );
+  }
+
+  getInnovationSections(innovationId: string): Observable<InnovationSectionsListDTO> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/sections')
+      .setPathParams({ innovationId });
+    return this.http.get<InnovationSectionsListDTO>(url.buildUrl()).pipe(
+      take(1),
+      map(response => response)
+    );
+  }
+
+  getSectionInfo(
+    innovationId: string,
+    sectionId: string,
+    filters: { fields?: 'tasks'[] }
+  ): Observable<InnovationSectionInfoDTO> {
+    const qp = {
+      ...(filters.fields ? { fields: filters.fields } : {})
+    };
+
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/sections/:sectionId')
+      .setPathParams({ innovationId, sectionId })
+      .setQueryParams(qp);
+    return this.http.get<InnovationSectionInfoDTO>(url.buildUrl()).pipe(
+      take(1),
+      map(response => response)
+    );
+  }
+
+  getAllSectionsInfo(innovationId: string): Observable<InnovationAllSectionsInfoDTO> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/all-sections')
+      .setPathParams({ innovationId });
+    return this.http.get<InnovationAllSectionsInfoDTO>(url.buildUrl()).pipe(
+      take(1),
+      map(response => response)
+    );
+  }
+
+  updateSectionInfo(innovationId: string, sectionKey: string, data: MappedObjectType): Observable<MappedObjectType> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/sections/:sectionKey')
+      .setPathParams({ innovationId, sectionKey });
+    return this.http.put<MappedObjectType>(url.buildUrl(), data).pipe(take(1));
+  }
+
+  submitSections(innovationId: string, sectionKey: string): Observable<any> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/sections/:sectionKey/submit')
+      .setPathParams({ innovationId, sectionKey });
+    return this.http.patch<any>(url.buildUrl(), { sectionKey }).pipe(take(1));
+  }
+
+  getSectionEvidenceInfo(innovationId: string, evidenceId: string): Observable<GetInnovationEvidenceDTO> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/evidences/:evidenceId')
+      .setPathParams({ innovationId, evidenceId });
+    return this.http.get<GetInnovationEvidenceDTO>(url.buildUrl()).pipe(take(1));
+  }
+
+  upsertSectionEvidenceInfo(
+    innovationId: string,
+    data: MappedObjectType,
+    evidenceId?: string
+  ): Observable<{ id: string }> {
+    if (evidenceId) {
+      const url = new UrlModel(this.API_INNOVATIONS_URL)
+        .addPath('v1/:innovationId/evidence/:evidenceId')
+        .setPathParams({ innovationId, evidenceId });
+      return this.http.put<void>(url.buildUrl(), { ...data }).pipe(
+        take(1),
+        map(() => ({ id: evidenceId }))
+      );
+    } else {
+      const url = new UrlModel(this.API_INNOVATIONS_URL)
+        .addPath('v1/:innovationId/evidence')
+        .setPathParams({ innovationId });
+      return this.http.post<{ id: string }>(url.buildUrl(), data).pipe(take(1));
+    }
+  }
+
+  deleteEvidence(innovationId: string, evidenceId: string): Observable<void> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/evidence/:evidenceId')
+      .setPathParams({ innovationId, evidenceId });
+    return this.http.delete<void>(url.buildUrl()).pipe(take(1));
+  }
+
+  getInnovationOrganisationSuggestions(
+    innovationId: string,
+    filters: { majorAssessmentId?: string }
+  ): Observable<OrganisationSuggestionModel> {
+    const qp = {
+      ...(filters.majorAssessmentId ? { majorAssessmentId: filters.majorAssessmentId } : {})
+    };
+
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/suggestions')
+      .setPathParams({ innovationId })
+      .setQueryParams(qp);
+
+    return this.http.get<OrganisationSuggestionModel>(url.buildUrl()).pipe(take(1));
+  }
+
+  getInnovationQASuggestions(innovationId: string): Observable<InnovationUnitSuggestionsType> {
+    const url = new UrlModel(this.API_INNOVATIONS_URL)
+      .addPath('v1/:innovationId/units-suggestions')
+      .setPathParams({ innovationId });
+    return this.http.get<InnovationUnitSuggestionsType>(url.buildUrl()).pipe(take(1));
   }
 }
